Update Identities intro page
