+++ packages/frontend/src/app/identities/page.js
@@ -2,8 +2,8 @@ import Identities from './Identities'
 import Intro from '../../components/intro'
 import Markdown from '../../components/markdown'
 import introContent from './intro.md'
-import Cards from './Cards'
-import { Container, Flex, Box } from '@chakra-ui/react'
+import { Container } from '@chakra-ui/react'
+import { TopIdentities } from '../../components/identities'
 
 export const metadata = {
   title: 'Identities — Dash Platform Explorer',
@@ -23,24 +23,11 @@ function IdentitiesRoute ({ searchParams }) {
       mt={8}
       mb={0}
     >
-      <Flex
-          justifyContent={'space-between'}
-          alignItems={'center'}
-          wrap={['wrap', 'wrap', 'wrap', 'nowrap']}
-      >
-          <Container maxW={['100%', '100%', '100%', 'calc(50% - 20px)']}>
-            <Intro
-              title={'Identities'}
-              contentSource={<Markdown>{introContent}</Markdown>}
-            />
-          </Container>
-
-          <Box flexShrink={'0'} w={10} h={[5, 5, 5, 10]}/>
-
-          <Container maxW={'none'} p={0}>
-            <Cards/>
-          </Container>
-      </Flex>
+      <Intro
+        title={'Identities'}
+        description={<Markdown>{introContent}</Markdown>}
+        block={<TopIdentities/>}
+      />
     </Container>
     <Identities defaultPage={page} defaultPageSize={pageSize}/>
   </>
