+++ packages/frontend/src/components/identities/TopIdentities.scss
@@ -1,0 +1,3 @@
+.TopIdentities {
+  width: 100%;
+}
