+++ packages/frontend/src/components/identities/index.js
@@ -1,10 +1,12 @@
 import RichList from './RichList'
 import IdentitiesList from './IdentitiesList'
 import { IdentitiesCards, IdentityCard } from './Cards'
+import TopIdentities from './TopIdentities'
 
 export {
   RichList,
   IdentitiesList,
   IdentitiesCards,
-  IdentityCard
+  IdentityCard,
+  TopIdentities
 }
